Engine fitting
Jet engines can now automatically fit their static thrust to give the
correct dry thrust in the event of plugin or config changes.

Results are stored in a database so it doesn't have to be done if
nothing has changed.

Will add sfc, wet thrust to fit other parameters in the future, however
those require nonlinear fitting techniques which are more complicated.